zk-voting: move helper functions to sdk
