+++ cli/package.json
@@ -109,14 +109,14 @@
     "test-psp-add-program": "cd ../psp-examples/tmp-test-psp  &&  ../../cli/test_bin/run psp:add-program --path \"../../psp-template\" --lspv \"{ path = '../../../../programs/psp4in4out-app-storage', features = ['cpi'] }\"  --m \"{ path = '../../../../macros/light' }\"  --lmtv \"{ path = '../../../../programs/merkle-tree', features = ['cpi'] }\" --vsdk \"{ path = '../../../../verifier-sdk'}\" added-psp && cd ../../cli && ts-node scripts/checkAddProgram.ts ../psp-examples/tmp-test-psp  added-psp",
     "test-psp-init": "cd ../psp-examples && rm -r tmp-test-psp || true && ../cli/test_bin/run psp:init --path \"../psp-template\" --zkjs \"workspace:*\" --pjs \"workspace:*\" --clib \"workspace:*\" --lspv \"{ path = '../../../../programs/psp4in4out-app-storage', features = ['cpi'] }\"  --m \"{ path = '../../../../macros/light' }\"  --lmtv \"{ path = '../../../../programs/merkle-tree', features = ['cpi'] }\" --vsdk \"{ path = '../../../../verifier-sdk'}\" tmp-test-psp && cd ../cli",
     "test-psp-add-circuit": "cd ../psp-examples/tmp-test-psp && ./../../cli/test_bin/run psp:add-circuit second-circuit --programName tmp-test-psp --path \"../../psp-template\"  --circom && ./../../cli/test_bin/run psp:add-circuit second-circuit1 --path \"../../psp-template\"  --programName tmp-test-psp  && cd ../../cli && ts-node scripts/checkAddCircuit.ts ../psp-examples/tmp-test-psp second-circuit tmp-test-psp true && ts-node scripts/checkAddCircuit.ts ../psp-examples/tmp-test-psp second-circuit1 tmp-test-psp",
-    "test-psp-build": "cd ../psp-examples/tmp-test-psp && pnpm i --no-lockfile && ./../../cli/test_bin/run psp:build tmp-test-psp && cd ../../cli",
-    "test-psp-test": "cd ../psp-examples/tmp-test-psp && ./../../cli/test_bin/run psp:test tmp-test-psp Fg6PaFpoGXkYsidMpWTK6W2BeZ7FEfcYkg476zPFsLnS && cd ../../cli",
+    "test-psp-build": "cd ../psp-examples/tmp-test-psp && pnpm i && ./../../cli/test_bin/run psp:build tmp-test-psp && cd ../../cli",
+    "test-psp-test": "cd ../psp-examples/tmp-test-psp && ./../../cli/test_bin/run psp:test tmp-test-psp && cd ../../cli",
     "clean-test-psp": "rm -r ../psp-examples/tmp-test-psp || true",
     "test-psp": "pnpm test-psp-init && pnpm test-psp-build && pnpm test-psp-test && pnpm test-psp-add-circuit",
     "test-circom-init": "cd ../psp-examples/ && rm -r tmp-test-circom || true && ./../cli/test_bin/run circom:init  tmp-test-circom --path \"../psp-template\"  && cd ../cli",
     "test-circom-add-circuit": "cd ../psp-examples/tmp-test-circom && ./../../cli/test_bin/run circom:add-circuit second-circuit --programName tmp-test-circom --path \"../../psp-template\"  && cd ../../cli && ts-node scripts/checkAddCircuit.ts ../psp-examples/tmp-test-circom second-circuit tmp-test-circom true",
-    "test-circom-build": "cd ../psp-examples/tmp-test-circom && pnpm i --no-lockfile && ./../../cli/test_bin/run circom:build tmp-test-circom && cd ../../cli",
-    "test-circom-test": "cd ../psp-examples/tmp-test-circom && ./../../cli/test_bin/run circom:test tmp-test-circom Fg6PaFpoGXkYsidMpWTK6W2BeZ7FEfcYkg476zPFsLnS && cd ../../cli",
+    "test-circom-build": "cd ../psp-examples/tmp-test-circom && pnpm i && ./../../cli/test_bin/run circom:build tmp-test-circom && cd ../../cli",
+    "test-circom-test": "cd ../psp-examples/tmp-test-circom && ./../../cli/test_bin/run circom:test tmp-test-circom && cd ../../cli",
     "clean-test-circom": "rm -r ../psp-examples/tmp-test-circom || true",
     "test-circom": "pnpm test-circom-init && pnpm test-circom-build && pnpm test-circom-test && pnpm test-circom-add-circuit",
     "test": "pnpm kill && pnpm test-cli && pnpm test-psp && pnpm test-circom && pnpm test-utils",
