+++ psp-examples/zk-voting/tests/init-vote-circuit.test.ts
@@ -1,18 +1,10 @@
 import * as anchor from "@coral-xyz/anchor";
-import { assert } from "chai";
-import {
-  TestRelayer,
-  Provider,
-  circuitlibjs,
-  Account,
-} from "@lightprotocol/zk.js";
-import { Prover } from "@lightprotocol/prover.js";
-const { MerkleTree, ElGamalUtils } = circuitlibjs;
+import { TestRelayer, Provider, circuitlibjs } from "@lightprotocol/zk.js";
+const { ElGamalUtils } = circuitlibjs;
 const { pointToStringArray, coordinatesToExtPoint } = ElGamalUtils;
 import { PublicKey, Keypair } from "@solana/web3.js";
 import {
   encrypt,
-  PublicKey as ElGamalPublicKey,
   generateKeypair,
   generateRandomSalt,
 } from "@lightprotocol/circuit-lib.js";
@@ -22,6 +14,8 @@ import { BN } from "@coral-xyz/anchor";
 import { IDL, PrivateVoting } from "../target/types/private_voting";
 import { utils } from "@project-serum/anchor";
 const path = require("path");
+
+import { InitVoteTransactionInput, createInitVoteProof } from "../sdk/index";
 
 const verifierProgramId = new PublicKey(
   "Fg6PaFpoGXkYsidMpWTK6W2BeZ7FEfcYkg476zPFsLnS"
@@ -36,7 +30,7 @@ const RPC_URL = "http://127.0.0.1:8899";
  * 3. init vote
  * 4. vote
  */
-describe("Test private-voting", () => {
+describe("Test Init Vote Circuit", () => {
   process.env.ANCHOR_PROVIDER_URL = RPC_URL;
   process.env.ANCHOR_WALLET = process.env.HOME + "/.config/solana/id.json";
 
@@ -125,41 +119,3 @@ describe("Test private-voting", () => {
     await createInitVoteProof(initVoteTransactionInput);
   });
 });
-
-export type InitVoteTransactionInput = {
-  idl: anchor.Idl;
-  elGamalPublicKey: ElGamalPublicKey;
-  circuitPath: string;
-};
-
-export const createInitVoteProof = async (
-  voteTransactionInput: InitVoteTransactionInput
-) => {
-  const { idl, circuitPath, elGamalPublicKey } = voteTransactionInput;
-  const yesZeroNonce = generateRandomSalt();
-  const { ephemeralKey: zeroYesEmphemeralKey, ciphertext: zeroYesCiphertext } =
-    encrypt(elGamalPublicKey, BigInt(0), yesZeroNonce);
-
-  const zeroCiphertextString = pointToStringArray(zeroYesCiphertext);
-  const zeroEmphemeralKeyString = pointToStringArray(zeroYesEmphemeralKey);
-  const elGamalPublicKeyString = pointToStringArray(elGamalPublicKey);
-
-  const publicInputs = {
-    publicElGamalPublicKeyX: new BN(elGamalPublicKeyString[0]),
-    publicElGamalPublicKeyY: new BN(elGamalPublicKeyString[1]),
-    publicZeroYesEmphemeralKeyX: new BN(zeroEmphemeralKeyString[0]),
-    publicZeroYesEmphemeralKeyY: new BN(zeroEmphemeralKeyString[1]),
-    publicZeroYesCiphertextX: new BN(zeroCiphertextString[0]),
-    publicZeroYesCiphertextY: new BN(zeroCiphertextString[1]),
-  };
-  const proofInputs = {
-    ...publicInputs,
-    nonce: new BN(yesZeroNonce.toString()),
-  };
-  const prover = new Prover(idl, circuitPath, "initVote");
-  await prover.addProofInputs(proofInputs);
-  console.time("Init vote proof: ");
-  const { parsedProof, parsedPublicInputs } = await prover.fullProveAndParse();
-  console.timeEnd("Init vote proof: ");
-  return { proof: parsedProof, publicInputs: parsedPublicInputs };
-};
